mips: correct the forgotten size_t address in the last API change

+++ arch/Mips/MipsDisassembler.c
@@ -296,7 +296,7 @@ bool Mips_getInstruction(csh ud, unsigned char *code, size_t code_len, MCInst *i
 static DecodeStatus Mips64Disassembler_getInstruction(int mode, MCInst *instr,
 		unsigned char *code, size_t code_len,
 		uint16_t *Size,
-		size_t Address, bool isBigEndian, MCRegisterInfo *MRI)
+		uint64_t Address, bool isBigEndian, MCRegisterInfo *MRI)
 {
 	uint32_t Insn;
 
@@ -321,7 +321,7 @@ static DecodeStatus Mips64Disassembler_getInstruction(int mode, MCInst *instr,
 }
 
 bool Mips64_getInstruction(csh ud, unsigned char *code, size_t code_len, MCInst *instr,
-		uint16_t *size, size_t address, void *info)
+		uint16_t *size, uint64_t address, void *info)
 {
 	cs_struct *handle = (cs_struct *)(uintptr_t)ud;
